+++ package.json
@@ -39,15 +39,19 @@
     "@koishijs/plugin-manager": "^1.7.3",
     "@koishijs/plugin-sandbox": "^1.1.0",
     "@koishijs/plugin-status": "^5.3.6",
+    "@types/md5": "^2.3.2",
     "@types/nanoid-dictionary": "^4.2.0",
+    "@types/node-fetch": "^2.6.1",
     "chrome-finder": "^1.0.7",
     "emoji-regex": "^10.1.0",
     "koishi": "^4.6.1",
+    "md5": "^2.3.0",
     "mongodb": "^4.5.0",
     "nanoid": "^3.3.3",
     "nanoid-dictionary": "^4.3.0",
     "node-fetch": "^3.2.4",
     "puppeteer-core": "^13.7.0",
+    "ts-retry-promise": "^0.6.1",
     "twemoji": "^14.0.2",
     "twitter-api-v2": "^1.12.0"
   }
