+++ plugins/baidu-translate/package.json
@@ -1,0 +1,22 @@
+{
+  "name": "koishi-plugin-baidu-translate",
+  "description": "translate given text using baidu translate api",
+  "private": true,
+  "version": "1.0.0",
+  "main": "lib/index.js",
+  "typings": "lib/index.d.ts",
+  "files": [
+    "lib",
+    "dist"
+  ],
+  "license": "MIT",
+  "scripts": {},
+  "keywords": [
+    "chatbot",
+    "koishi",
+    "plugin"
+  ],
+  "peerDependencies": {
+    "koishi": "^4.6.1"
+  }
+}
