+++ plugins/baidu-translate/readme.md
@@ -1,0 +1,5 @@
+# koishi-plugin-baidu-translate
+
+[![npm](https://img.shields.io/npm/v/koishi-plugin-baidu-translate?style=flat-square)](https://www.npmjs.com/package/koishi-plugin-baidu-translate)
+
+translate given text using baidu translate api
