+++ plugins/baidu-translate/tsconfig.json
