+++ plugins/mongo-database/readme.md
@@ -1,0 +1,5 @@
+# koishi-plugin-mongo-database
+
+[![npm](https://img.shields.io/npm/v/koishi-plugin-mongo-database?style=flat-square)](https://www.npmjs.com/package/koishi-plugin-mongo-database)
+
+custom database specifically designed for this bot
