+++ plugins/mongo-database/src/model.ts
@@ -1,28 +1,49 @@
+/**
+ * This interface represents the user config for a twitter user
+ */
 export interface IUserConfig {
   readonly userid: string,
   readonly username: string,
+  // receive streamed tweet based on type
   tweet: boolean,
   retweet: boolean,
   comment: boolean,
+  // receive streamed tweet content
   text: boolean,
   screenshot: boolean,
   translation: boolean,
   extended: boolean,
+  // not implemented
   forwardMsg: boolean,
-  css? :string,
+  // custom translate content
+  css?: string,
   tag?: string,
   background?: string,
 }
 
+// hard coded max length of history, ensures that the database will not become too huge
+export const MAX_HISTORY_LENGTH = 200;
+
+/**
+ * This interface represents the group config for a group
+ */
+export interface IGroupConfig {
+  readonly guildId: string,
+  userConfigMap: { [userid: string]: IUserConfig },
+  historyList: string[],
+  // monotonically increase index
+  currentIndex: number,
+}
+
 type TestEqual<T, U, Y=true, N=false> = (<G>() => G extends T ? 1 : 2) extends (<G>() => G extends U ? 1 : 2) ? Y : N;
 
-export const InvalidUserConfig = ["userId", "username"] as const;
+export const InvalidUserConfig = ["userid", "username"] as const;
 
 type MutableUserConfig = Omit<IUserConfig, (typeof InvalidUserConfig)[number]>;
 
 type SwitchableUserConfig = {[Key in keyof MutableUserConfig]-?: MutableUserConfig[Key] extends boolean ? Key : never}[keyof MutableUserConfig];
 // CHECK EQUALITY BELOW!!
-export const SwitchableUserConfigKeys = ["comment", "content", "forwardMsg", "retweet", "screenshot", "text", "translation", "tweet"] as const;
+export const SwitchableUserConfigKeys = ["comment", "extended", "forwardMsg", "retweet", "screenshot", "text", "translation", "tweet"] as const;
 type TestSwitchableEquality = TestEqual<SwitchableUserConfig, (typeof SwitchableUserConfigKeys)[number]>;
 
 type CustomizableUserConfig = {[Key in keyof MutableUserConfig]-?: MutableUserConfig[Key] extends boolean ? never : Key}[keyof MutableUserConfig];
@@ -35,11 +56,18 @@ export type ImmutableUserConfigProperties = (typeof immutableUserConfigPropertie
 
 export type UserConfigModifier = Partial<Omit<IUserConfig, ImmutableUserConfigProperties>>;
 
-export const MAX_HISTORY_LENGTH = 200;
+/**
+ * This interface represents a succeed execution
+ */
+export interface Ok<T> {
+  state: true,
+  content: T,
+}
 
-export interface IGroupConfig {
-  readonly guildId: string,
-  userConfigMap: {[userid: string]: IUserConfig},
-  historyList: string[]
-  currentIndex: number,
+/**
+ * This interface represents a failed execution
+ */
+export interface Err<T> {
+  state: false,
+  content: T,
 }
