+++ plugins/twitter-api-client/readme.md
@@ -1,0 +1,5 @@
+# koishi-plugin-twitter-api-client
+
+[![npm](https://img.shields.io/npm/v/koishi-plugin-twitter-api-client?style=flat-square)](https://www.npmjs.com/package/koishi-plugin-twitter-api-client)
+
+twitter api v2 client and stream client
