+++ plugins/twitter-command/readme.md
@@ -1,0 +1,5 @@
+# koishi-plugin-twitter-command
+
+[![npm](https://img.shields.io/npm/v/koishi-plugin-twitter-command?style=flat-square)](https://www.npmjs.com/package/koishi-plugin-twitter-command)
+
+commands of twitter plugin
