+++ plugins/twitter-screenshot-client/readme.md
@@ -1,0 +1,5 @@
+# koishi-plugin-twitter-screenshot-client
+
+[![npm](https://img.shields.io/npm/v/koishi-plugin-twitter-screenshot-client?style=flat-square)](https://www.npmjs.com/package/koishi-plugin-twitter-screenshot-client)
+
+take screenshot for twitter webpage
