--- plugins/twitter/readme.md
@@ -1,5 +1,0 @@
-# koishi-plugin-twitter
-
-[![npm](https://img.shields.io/npm/v/koishi-plugin-twitter?style=flat-square)](https://www.npmjs.com/package/koishi-plugin-twitter)
-
-take screenshot of tweets and send to groups
