--- plugins/twitter/src/mongodatabase/index.ts
@@ -1,2 +1,0 @@
-export * from "./client";
-export * from "./model";
