--- plugins/twitter/src/twitter-api/index.ts
@@ -1,1 +1,0 @@
-export * from "./client";
